requirements.txt
uvicorn
fastapi
pillow
python-multipart
